likes

+++ app.py
@@ -23,7 +23,10 @@ db = SQL("sqlite:///database.db")
 @login_required
 def home():
     if request.method == "POST":
-        db.execute("DELETE FROM images WHERE id =  ?", request.form.get("admin"))
+        if request.form.get("admin"):
+            db.execute("DELETE FROM images WHERE id =  ?", request.form.get("admin"))
+        if request.form.get("like"):
+            db.execute("UPDATE images SET likes = likes + 1 WHERE id = ?", request.form.get("like"))
     if session["user_id"] == 1:
         admin = 1
     else:
